Lehao - Undo command bug fixes
The previous undo has such a bug: when delete person and then undo, the client list of the affected appointment will not be refreshed (undo). This is because a state simply copy the reference of the schedule, and different schedules in different states actually refer to the same appointment and same client list. To solve this we need to create deep copy when initialize the state and get schedule of the state.